+++ RefactorThis.Persistence/Entities/Invoice.cs
@@ -1,32 +1,14 @@
 using System.Collections.Generic;
-using RefactorThis.Persistence.Repositories;
+using RefactorThis.Persistence.Enums;
 
 namespace RefactorThis.Persistence.Entities
 {
     public class Invoice
     {
-        private readonly InvoiceRepository _repository;
-        public Invoice(InvoiceRepository repository)
-        {
-            _repository = repository;
-        }
-
-        public void Save()
-        {
-            _repository.SaveInvoice(this);
-        }
-
         public decimal Amount { get; set; }
         public decimal AmountPaid { get; set; }
         public decimal TaxAmount { get; set; }
         public List<Payment> Payments { get; set; }
-
         public InvoiceType Type { get; set; }
-    }
-
-    public enum InvoiceType
-    {
-        Standard,
-        Commercial
     }
 }
