+++ RefactorThis.Persistence/Enums/InvoiceType.cs
@@ -1,0 +1,8 @@
+﻿namespace RefactorThis.Persistence.Enums
+{
+    public enum InvoiceType
+    {
+        Standard,
+        Commercial
+    }
+}
